CPO-6: refactor after merge with master
   [CPO-6](https://tools.hmcts.net/jira/browse/CPO-6)
   [CPO-7](https://tools.hmcts.net/jira/browse/CPO-7)

Refactor CPO-6 after merger with master including CPO-7 code:
* remove unused mapper function.
* adjust for some Sonar warnings.
* refactor error handlers.
* consolidate CPO comptroller test classes into 1 from 2
* consolidate CPO service test classes into 1 from 3
