+++ src/main/java/uk/gov/hmcts/reform/cpo/service/mapper/CasePaymentOrderMapper.java
@@ -8,8 +8,6 @@ import uk.gov.hmcts.reform.cpo.domain.CasePaymentOrder;
 import uk.gov.hmcts.reform.cpo.payload.CreateCasePaymentOrderRequest;
 import uk.gov.hmcts.reform.cpo.payload.UpdateCasePaymentOrderRequest;
 
-import java.util.List;
-
 @Mapper(componentModel = "spring")
 public interface CasePaymentOrderMapper {
 
@@ -21,8 +19,6 @@ public interface CasePaymentOrderMapper {
                                                 String createdBy);
 
     CasePaymentOrder toDomainModel(CasePaymentOrderEntity casePaymentOrderEntity);
-
-    List<CasePaymentOrder> toDomainModelList(List<CasePaymentOrderEntity> casePaymentOrderEntities);
 
     @Mapping(target = "createdTimestamp", expression = "java(java.time.LocalDateTime.now())")
     void mergeIntoEntity(@MappingTarget CasePaymentOrderEntity target,
