+++ src/test/java/uk/gov/hmcts/reform/BaseTest.java
@@ -58,8 +58,6 @@ public interface BaseTest {
     String CREATED_BY = "createdBy";
     LocalDateTime CREATED_TIMESTAMP = LocalDateTime.now();
 
-    String UN_EXPECTED_ERROR_IN_TEST = "Fail due an expected error in the test.";
-
     default <T> Optional<List<T>> createInitialValuesList(final T[] initialValues) {
         return Optional.of(Arrays.asList(initialValues));
     }
@@ -177,8 +175,8 @@ public interface BaseTest {
     default void assertGetCopPResponse(String expectedError, ResultActions response) throws Exception {
         response
             .andExpect(status().isBadRequest())
-            .andExpect(jsonPath("$.status").value(HttpStatus.BAD_REQUEST.value()))
-            .andExpect(jsonPath("$.error").value(HttpStatus.BAD_REQUEST.getReasonPhrase()))
-            .andExpect(jsonPath("$.message",containsString(expectedError)));
+            .andExpect(jsonPath(ERROR_PATH_STATUS).value(HttpStatus.BAD_REQUEST.value()))
+            .andExpect(jsonPath(ERROR_PATH_ERROR).value(HttpStatus.BAD_REQUEST.getReasonPhrase()))
+            .andExpect(jsonPath(ERROR_PATH_MESSAGE, containsString(expectedError)));
     }
 }
