+++ src/test/java/uk/gov/hmcts/reform/cpo/service/impl/CasePaymentOrdersServiceImplTest.java
@@ -5,10 +5,17 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.mockito.ArgumentCaptor;
+import org.mockito.ArgumentMatchers;
+import org.mockito.Captor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import uk.gov.hmcts.reform.BaseTest;
 import uk.gov.hmcts.reform.cpo.data.CasePaymentOrderEntity;
 import uk.gov.hmcts.reform.cpo.domain.CasePaymentOrder;
@@ -17,20 +24,31 @@ import uk.gov.hmcts.reform.cpo.exception.CasePaymentOrderCouldNotBeFoundExceptio
 import uk.gov.hmcts.reform.cpo.exception.IdAMIdCannotBeRetrievedException;
 import uk.gov.hmcts.reform.cpo.payload.CreateCasePaymentOrderRequest;
 import uk.gov.hmcts.reform.cpo.payload.UpdateCasePaymentOrderRequest;
+import uk.gov.hmcts.reform.cpo.repository.CasePaymentOrderQueryFilter;
 import uk.gov.hmcts.reform.cpo.repository.CasePaymentOrdersRepository;
 import uk.gov.hmcts.reform.cpo.security.SecurityUtils;
 import uk.gov.hmcts.reform.cpo.service.mapper.CasePaymentOrderMapperImpl;
 import uk.gov.hmcts.reform.cpo.validators.ValidationError;
 import uk.gov.hmcts.reform.idam.client.models.UserInfo;
 
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+import static org.springframework.test.util.AssertionErrors.assertTrue;
 import static uk.gov.hmcts.reform.cpo.data.CasePaymentOrderEntity.UNIQUE_CASE_ID_ORDER_REF_CONSTRAINT;
 import static uk.gov.hmcts.reform.cpo.validators.ValidationError.CASE_ID_ORDER_REFERENCE_UNIQUE;
 import static uk.gov.hmcts.reform.cpo.validators.ValidationError.IDAM_ID_RETRIEVE_ERROR;
@@ -48,6 +66,12 @@ class CasePaymentOrdersServiceImplTest implements BaseTest {
 
     @Mock
     private SecurityUtils securityUtils;
+
+    @Captor
+    ArgumentCaptor<List<UUID>> uuidArgumentCaptor;
+
+    @Captor
+    ArgumentCaptor<List<Long>> caseIdsArgumentCaptor;
 
     @BeforeEach
     public void setUp() {
@@ -116,6 +140,169 @@ class CasePaymentOrdersServiceImplTest implements BaseTest {
                 .hasMessageContaining(IDAM_ID_RETRIEVE_ERROR);
         }
     }
+
+
+    @Nested
+    @DisplayName("Delete Case Payment Order")
+    class DeleteCasePaymentOrder {
+
+        private final List<UUID> uuidsToDelete = List.of(UUID.randomUUID(), UUID.randomUUID());
+
+        private final List<Long> caseIdsToDelete = List.of(123L, 456L);
+
+        private CasePaymentOrderQueryFilter uuidFilter;
+
+        private CasePaymentOrderQueryFilter caseIdFilter;
+
+        @BeforeEach
+        void beforeEachTest() {
+            uuidFilter = CasePaymentOrderQueryFilter.builder()
+                .cpoIds(uuidsToDelete.stream()
+                            .map(UUID::toString)
+                            .collect(Collectors.toList()))
+                .caseIds(Collections.emptyList())
+                .build();
+            caseIdFilter = CasePaymentOrderQueryFilter.builder()
+                .cpoIds(Collections.emptyList())
+                .caseIds(caseIdsToDelete.stream()
+                             .map(Object::toString)
+                             .collect(Collectors.toList()))
+                .build();
+        }
+
+        @Test
+        void deleteCasePaymentOrdersById() {
+            casePaymentOrdersService.deleteCasePaymentOrders(uuidFilter);
+
+            verify(casePaymentOrdersRepository).deleteByUuids(uuidArgumentCaptor.capture());
+            assertEquals(uuidsToDelete, uuidArgumentCaptor.getValue());
+
+            verify(casePaymentOrdersRepository).deleteAuditEntriesByUuids(uuidArgumentCaptor.capture());
+            assertEquals(uuidsToDelete, uuidArgumentCaptor.getValue());
+        }
+
+        @Test
+        void deleteCasePaymentOrdersByCaseIds() {
+            casePaymentOrdersService.deleteCasePaymentOrders(caseIdFilter);
+
+            verify(casePaymentOrdersRepository).deleteByCaseIds(caseIdsArgumentCaptor.capture());
+            assertEquals(caseIdsArgumentCaptor.getValue(), caseIdsToDelete);
+
+            verify(casePaymentOrdersRepository).deleteAuditEntriesByCaseIds(caseIdsArgumentCaptor.capture());
+            assertEquals(caseIdsArgumentCaptor.getValue(), caseIdsToDelete);
+        }
+    }
+
+
+    @Nested
+    @DisplayName("Get Case Payment Order")
+    class GetCasePaymentOrder {
+
+        private final List<String> casesIds = List.of("1609243447569251", "1609243447569252", "1609243447569253");
+
+        private final List<String> ids = List.of("df54651b-3227-4067-9f23-6ffb32e2c6bd",
+                                                 "d702ef36-0ca7-46e9-8a00-ef044d78453e",
+                                                 "d702ef36-0ca7-46e9-8a00-ef044d78453e");
+
+        @Test
+        void passForListIds() {
+            final CasePaymentOrderQueryFilter casePaymentOrderQueryFilter = CasePaymentOrderQueryFilter.builder()
+                .cpoIds(ids)
+                .caseIds(Collections.emptyList())
+                .pageable(getPageRequest())
+                .build();
+
+            when(casePaymentOrdersRepository.findByIdIn(anyList(), ArgumentMatchers.<Pageable>any())).thenReturn(
+                getEntityPages());
+
+            final Page<CasePaymentOrder> pages = casePaymentOrdersService.getCasePaymentOrders(
+                casePaymentOrderQueryFilter);
+
+            assertTrue("The getNumberOfElements should be 0.", pages.getNumberOfElements() == 3);
+        }
+
+        @Test
+        void passForListCasesIds() {
+            final CasePaymentOrderQueryFilter casePaymentOrderQueryFilter = CasePaymentOrderQueryFilter.builder()
+                .cpoIds(Collections.emptyList())
+                .caseIds(casesIds)
+                .pageable(getPageRequest())
+                .build();
+
+            when(casePaymentOrdersRepository.findByCaseIdIn(anyList(), ArgumentMatchers.<Pageable>any())).thenReturn(
+                getEntityPages());
+
+            final Page<CasePaymentOrder> pages = casePaymentOrdersService.getCasePaymentOrders(
+                casePaymentOrderQueryFilter);
+
+            assertTrue("The getNumberOfElements should be 0.", pages.getNumberOfElements() == 3);
+        }
+
+        @Test
+        void failForListCasesIds() {
+            final ArrayList<CasePaymentOrderEntity> casePaymentOrders = new ArrayList<>();
+            final Page<CasePaymentOrderEntity> pageImpl = new PageImpl<>(
+                casePaymentOrders,
+                getPageRequest(),
+                3
+            );
+            final CasePaymentOrderQueryFilter casePaymentOrderQueryFilter = CasePaymentOrderQueryFilter.builder()
+                .cpoIds(Collections.emptyList())
+                .caseIds(casesIds)
+                .pageable(getPageRequest())
+                .build();
+
+            when(casePaymentOrdersRepository.findByCaseIdIn(anyList(), any())).thenReturn(
+                pageImpl);
+
+            assertThatThrownBy(() -> casePaymentOrdersService.getCasePaymentOrders(casePaymentOrderQueryFilter))
+                .isInstanceOf(CasePaymentOrderCouldNotBeFoundException.class)
+                .hasMessageContaining(ValidationError.CPO_NOT_FOUND);
+        }
+
+        private Page<CasePaymentOrderEntity> getEntityPages() {
+            final PageRequest pageRequest = getPageRequest();
+            return new PageImpl<CasePaymentOrderEntity>(createListOfCasePaymentOrderEntity(), pageRequest, 3);
+        }
+
+
+        private List<CasePaymentOrderEntity> createListOfCasePaymentOrderEntity() {
+            final ArrayList<CasePaymentOrderEntity> casePaymentOrders = new ArrayList<>();
+
+            final CasePaymentOrderEntity casePaymentOrderEntity = new CasePaymentOrderEntity();
+            casePaymentOrderEntity.setAction("action");
+            casePaymentOrderEntity.setCaseId(Long.parseLong("1609243447569251"));
+            casePaymentOrderEntity.setCreatedBy("action1");
+            casePaymentOrderEntity.setOrderReference("action1");
+            casePaymentOrderEntity.setEffectiveFrom(LocalDateTime.now());
+            casePaymentOrderEntity.setCreatedTimestamp(LocalDateTime.now());
+            casePaymentOrderEntity.setResponsibleParty("setResponsibleParty");
+            casePaymentOrders.add(casePaymentOrderEntity);
+
+            final CasePaymentOrderEntity casePaymentOrderEntity1 = new CasePaymentOrderEntity();
+            casePaymentOrderEntity1.setAction("action");
+            casePaymentOrderEntity1.setCaseId(Long.parseLong("1609243447569252"));
+            casePaymentOrderEntity1.setCreatedBy("action1");
+            casePaymentOrderEntity1.setOrderReference("Baction2");
+            casePaymentOrderEntity1.setEffectiveFrom(LocalDateTime.now());
+            casePaymentOrderEntity1.setCreatedTimestamp(LocalDateTime.now());
+            casePaymentOrderEntity1.setResponsibleParty("setResponsibleParty");
+            casePaymentOrders.add(casePaymentOrderEntity1);
+
+            final CasePaymentOrderEntity casePaymentOrderEntity2 = new CasePaymentOrderEntity();
+            casePaymentOrderEntity2.setAction("action");
+            casePaymentOrderEntity2.setCaseId(Long.parseLong("1609243447569253"));
+            casePaymentOrderEntity2.setCreatedBy("action1");
+            casePaymentOrderEntity2.setOrderReference("Caction3");
+            casePaymentOrderEntity2.setEffectiveFrom(LocalDateTime.now());
+            casePaymentOrderEntity2.setCreatedTimestamp(LocalDateTime.now());
+            casePaymentOrderEntity2.setResponsibleParty("setResponsibleParty");
+            casePaymentOrders.add(casePaymentOrderEntity2);
+            return casePaymentOrders;
+        }
+
+    }
+
 
     @Nested
     @DisplayName("Update Case Payment Order")
@@ -254,6 +441,7 @@ class CasePaymentOrdersServiceImplTest implements BaseTest {
         }
 
     }
+
 
     private void setupSecurityUtilsMock() {
         UserInfo userInfo = UserInfo.builder()
